requirements.txt
streamlit~=1.43.2
pandas~=2.2.3
plotly~=6.0.0
pillow~=11.1.0
neo4j~=5.28.1
numpy~=2.2.3
networkx~=3.4.2